pyproject.toml
[project]
name = "Plasmid_LIN"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter-contrib-nbextensions>=0.7.0",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "mlflow>=3.8.1",
    "numpy>=2.4.0",
    "optuna>=4.6.0",
    "optuna-integration>=4.6.0",
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "seaborn>=0.12.0",
    "biopython>=1.84",
    "statsmodels>=0.14.6",
    "xgboost>=3.1.2",
]
